backend/analytics/serializers.py
from rest_framework import serializers
from .models import Recommendation


class RecommendationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Recommendation
        fields = ['id', 'title', 'description', 'recommendation_type', 'priority', 'is_dismissed', 'created_at']
